removed openapi blueprint and put everything under /swagger/, much cleaner solution which also solves the path to the spec.json error
