+++ examples/cars/main.py
@@ -1,5 +1,5 @@
 from sanic import Sanic
-from sanic_openapi import swagger_blueprint, openapi_blueprint
+from sanic_openapi import swagger_blueprint
 from blueprints.car import blueprint as car_blueprint
 from blueprints.driver import blueprint as driver_blueprint
 from blueprints.garage import blueprint as garage_blueprint
@@ -7,7 +7,6 @@ from blueprints.manufacturer import blueprint as manufacturer_blueprint
 
 app = Sanic()
 
-app.blueprint(openapi_blueprint)
 app.blueprint(swagger_blueprint)
 app.blueprint(car_blueprint)
 app.blueprint(driver_blueprint)
