+++ sanic_openapi/__init__.py
@@ -1,5 +1,4 @@
-from .openapi import blueprint as openapi_blueprint
 from .swagger import blueprint as swagger_blueprint
 
 __version__ = '0.5.0'
-__all__ = ['openapi_blueprint', 'swagger_blueprint']
+__all__ = ['swagger_blueprint']
